multi_agent_system/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# _libgcc_mutex==0.1
# _openmp_mutex==4.5
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asttokens
attrs
dataclasses-json
debugpy
decorator
distro
et-xmlfile
exceptiongroup
executing
fastapi
ffmpy
filelock
fonttools
frozenlist
fsspec
gradio
gradio-client
greenlet
h11
httpcore
httpx
idna
importlib-metadata
ipykernel
ipython
isort
jedi
jinja2
jsonpatch
jsonpointer
kiwisolver
krb5
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-experimental==0.3.2
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.136
matplotlib
matplotlib-inline
multidict
mypy-extensions

numpy
ollama
openai
openpyxl
pandas
pillow
prompt-toolkit
pydantic
pydantic-core
pydantic-settings
pydub
pygments
pyparsing
pyyaml

rdkit
readline
requests
requests-toolbelt
semantic-version
setuptools

six
sniffio
stack_data
starlette
tenacity
tk
tomlkit
tornado
traitlets
typer
typing-extensions
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
wcwidth
websockets
wheel
yarl
